Edit to inform user of nil quantity input errors

+++ app/controllers/meow/shopping_carts_controller.rb
@@ -4,9 +4,7 @@ class Meow::ShoppingCartsController < ApplicationController
     @shopping_cart = current_cart
     @line_items = @shopping_cart.line_items
     @line_items.each do |line_item|
-      if line_item.quantity == 0
-        line_item.destroy
-      end
+      line_item.destroy if line_item.quantity == 0
     end
   end
